UPD: categories/del.php - alerta ao excluir categoria vinculada a produto

+++ app/categories/del.php
@@ -6,18 +6,29 @@
     $link = "main.php?folder=categories/&file=frmins.php";
     $msg = '';
     $status = "success";
-    
-    $sql = "DELETE FROM categorias WHERE id = :id";
-    
+
+    $sql = "SELECT modelo FROM produtos WHERE categorias_id = :id";
+
     $stm_sql = $db_connection -> prepare($sql);
     $stm_sql -> bindParam(':id', $id);
-    $result = $stm_sql -> execute();
+    $stm_sql -> execute();
 
-    if ($result) {
-        $msg = "Categoria excluída com sucesso!";
+    if ($stm_sql -> rowCount() == 0) {
+        $sql = "DELETE FROM categorias WHERE id = :id";
+    
+        $stm_sql = $db_connection -> prepare($sql);
+        $stm_sql -> bindParam(':id', $id);
+        $result = $stm_sql -> execute();
+
+        if ($result) {
+            $msg = "Categoria excluída com sucesso!";
+        } else {
+            $msg = "Falha ao excluir categoria!";
+            $status = "danger";
+        }
     } else {
-        $msg = "Falha ao excluir categoria!";
-        $status = "danger";
+        $msg = "Existem produtos vinculados à esta categoria.";
+        $status = "warning";
     }
     header("Location: " . $link . "&mensagem=" . $msg . "&status=" . $status);
 ?>
